Struct dslab_dag::dag::DAG

pub struct DAG { /* private fields */ }

Represents a computation consisting of multiple tasks with data dependencies modeled as a directed acyclic graph (DAG).

Each task can produce one or more data items (task outputs) and consume (as task inputs) data items produced by other tasks. Entry tasks consume separate data items corresponding to the DAG inputs. The data dependencies between the tasks define constraints on task execution - a task cannot start its execution on some resource until all its inputs are produced (parent tasks are completed) and transferred to this resource.

Implementations

impl DAG

pub fn new() -> Self

Creates empty DAG.

pub fn from_file<P: AsRef<Path>>(file: P, parser_config: &ParserConfig) -> Self

Reads DAG from file in one of supported formats:

YAML format (.yaml extension)

WfCommons format (.json extension)

DAX format (.xml extension)

DOT format (.dot extension)

pub fn add_task( &mut self, name: &str, flops: f64, memory: u64, min_cores: u32, max_cores: u32, cores_dependency: CoresDependency ) -> usize

Adds new task with provided parameters and returns its id.

pub fn get_task(&self, task_id: usize) -> &Task

Returns task by id.

pub fn get_task_mut(&mut self, task_id: usize) -> &mut Task

Returns mutable task reference by id.

pub fn get_tasks(&self) -> &Vec<Task>

Returns all tasks.

pub fn get_data_item(&self, data_id: usize) -> &DataItem

Returns data item by id.

pub fn get_data_items(&self) -> &Vec<DataItem>

Returns all data items.

pub fn get_ready_tasks(&self) -> &BTreeSet<usize>

Returns ids of ready tasks.

pub fn get_inputs(&self) -> &BTreeSet<usize>

pub fn get_outputs(&self) -> &BTreeSet<usize>

pub fn add_data_item(&mut self, name: &str, size: f64) -> usize

Adds data item with provided parameters and returns its id.

pub fn add_task_output( &mut self, producer: usize, name: &str, size: f64 ) -> usize

Adds data item as a task output and returns its id.

pub fn add_data_dependency(&mut self, data_item_id: usize, consumer_id: usize)

Adds a dependency between data item and task.

pub fn update_task_state(&mut self, task_id: usize, state: TaskState)

Updates task state to a provided value, updating dependent data item states if needed.

pub fn update_data_item_state(&mut self, data_id: usize, state: DataItemState)

Updates data item state to a provided value, updating dependent task states if needed.

pub fn is_completed(&self) -> bool

Checks whether all tasks are completed.

pub fn stats(&self) -> DagStats

pub fn set_resource_restriction( &mut self, task_id: usize, restriction: ResourceRestriction )

impl DAG

pub fn from_dax<P: AsRef<Path>>(file: P, config: &ParserConfig) -> Self

Reads DAG from a file in DAX format.

impl DAG

pub fn from_dot<P: AsRef<Path>>(file: P) -> Self

Reads DAG from a file in DOT format.

impl DAG

pub fn from_wfcommons<P: AsRef<Path>>(file: P, config: &ParserConfig) -> Self

Reads DAG from a file in WfCommons json format.

impl DAG

pub fn from_yaml<P: AsRef<Path>>(file: P, config: &ParserConfig) -> Self

Reads DAG from a file in YAML format.