Struct dslab_dag::dag_simulation::DagSimulation

pub struct DagSimulation {
    pub sim: Simulation,
    /* private fields */
}

Provides a convenient API for configuring and running simulations of DAG execution.

Fields
sim: Simulation
Implementations

impl DagSimulation

pub fn new( seed: u64, resources: Vec<ResourceConfig>, network_config: NetworkConfig, scheduler: Rc<RefCell<dyn Scheduler>>, config: Config ) -> Self

Creates simulation with provided random seed, network model, scheduler and config.

pub fn add_resource(&mut self, name: &str, speed: f64, cores: u32, memory: u64)

Adds a resource with provided parameters.

pub fn init(&mut self, dag: DAG) -> Rc<RefCell<DAGRunner>>

Initializes DAG simulation.

pub fn steps(&mut self, step_count: u64) -> bool

Performs the specified number of steps through the simulation.

See Simulation::steps().

pub fn step_for_duration(&mut self, time: f64)

Steps through the simulation with duration limit.

See Simulation::step_for_duration().

pub fn step_until_no_events(&mut self)

Steps through the simulation until there are no pending events left.

See Simulation::step_until_no_events().

pub fn event_count(&self) -> u64

Returns the total number of created events.

See Simulation::event_count().

pub fn time(&mut self) -> f64

Returns the current simulation time.

See Simulation::time().