Trait dslab_dag::scheduler::Scheduler

pub trait Scheduler {
    // Required methods
    fn start(
        &mut self,
        dag: &DAG,
        system: System<'_>,
        config: Config,
        ctx: &SimulationContext
    ) -> Vec<Action>;
    fn is_static(&self) -> bool;

    // Provided method
    fn on_task_state_changed(
        &mut self,
        _task: usize,
        _task_state: TaskState,
        _dag: &DAG,
        _system: System<'_>,
        _ctx: &SimulationContext
    ) -> Vec<Action> { ... }
}

Trait for implementing DAG scheduling algorithms.

Includes callback methods which can return one or multiple actions corresponding to decisions made by the scheduler (assign task to resource, transfer data item between resources, etc).

Required Methods

fn start( &mut self, dag: &DAG, system: System<'_>, config: Config, ctx: &SimulationContext ) -> Vec<Action>

Called once in the beginning of DAG execution.

fn is_static(&self) -> bool

Should be true iff on_task_state_chaged always returns empty vector.

Provided Methods

fn on_task_state_changed( &mut self, _task: usize, _task_state: TaskState, _dag: &DAG, _system: System<'_>, _ctx: &SimulationContext ) -> Vec<Action>

Called on every task state change.

Useful for implementing dynamic scheduling algorithms. For static algorithms just return Vec::new().