Crate dslab_dag

DSLab DAG

A library for studying the scheduling of computations represented as directed acyclic graphs (DAG), such as scientific workflows or data-parallel jobs, in distributed systems. It allows to describe a computational DAG and simulate its execution in a given distributed system using the specified scheduling algorithm (including the user-defined one).

The distributed system is modeled as a set of computing resources connected with network. Each resource is described by the number of CPU cores, their speed in flop/s and amount of memory. Resources can execute compute tasks described by the amount of computations in flops, the minimum and maximum number of used cores, and the amount of used memory. Each task also has a function which defines the dependence of parallel speedup on the number of used cores. Currently, the allocation of cores for each task is computed upon the task arrival and is not changed during the task execution. The resource implementation is based on the multicore compute model from the dslab-compute crate. The network model is provided by the dslab-network crate.

The computational DAG is modeled as a set of tasks with data dependencies. Each task can produce one or more data items (task outputs) and consume (as task inputs) data items produced by other tasks. Entry tasks consume separate data items corresponding to the DAG inputs. The data dependencies between the tasks define constraints on task execution - a task cannot start its execution on some resource until all its inputs are produced (parent tasks are completed) and transferred to this resource.

The scheduling of DAG tasks, i.e. assigning tasks to resources and defining the order of task execution on each resource, is performed by the implementation of Scheduler trait. This trait includes two callback methods which are called at the start of DAG execution and on every task state change respectively. Each method can return one or multiple actions corresponding to decisions made by the scheduler (assign task to resource, transfer data item between resources, etc). This approach allows to implement and test arbitrary static or dynamic scheduling algorithms. The library includes several ready-to-use implementations in the schedulers folder.

DagSimulation provides a convenient API for configuring and running simulations. Here is a small example:

use std::rc::Rc;
use std::cell::RefCell;

use dslab_dag::dag::DAG;
use dslab_dag::dag_simulation::DagSimulation;
use dslab_dag::data_item::DataTransferMode;
use dslab_dag::network::read_network_config;
use dslab_dag::parsers::config::ParserConfig;
use dslab_dag::resource::read_resource_configs;
use dslab_dag::runner::Config;
use dslab_dag::schedulers::simple_scheduler::SimpleScheduler;

// load resources and network configuration
let resources = read_resource_configs("../../examples/dag-demo/systems/cluster-het-4-32cores.yaml");
let network = read_network_config("../../examples/dag-demo/systems/cluster-het-4-32cores.yaml");
// use simple scheduler implementation
let scheduler = Rc::new(RefCell::new(SimpleScheduler::new()));
// create simulation with random seed 123
let mut sim = DagSimulation::new(123, resources, network, scheduler, Config { data_transfer_mode: DataTransferMode::Direct });
// read DAG from YAML file
let dag = DAG::from_yaml("../../examples/dag-demo/dags/diamond.yaml", &ParserConfig::default());

// init simulation
let runner = sim.init(dag);
// run simulation until completion
sim.step_until_no_events();
// check that all tasks in DAG are completed
runner.borrow().validate_completed();

More examples can be found here in folders with “dag” prefix.

Modules

dag

DAG model of computation.

dag_simulation

Simulation configuration and execution.

dag_stats

data_item

Data item.

experiment

Tool for running experiments across many (dag, system. scheduler) combinations.

lower_bound

network

Network model tools.

parsers

Tools for loading DAG from different file formats.

resource

Resource model.

run_stats

runner

DAG execution runtime.

scheduler

DAG scheduling.

schedulers

Scheduler implementations.

system

Distributed computing system.

task

DAG task.

trace_log

DAG execution log.